move JsonSchemaStandaloneSource to top level class, move JsonSchemaToTypingResultConverter
